requirements_DS_train.txt
accelerate==0.28.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.6.2.post1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
bitsandbytes==0.44.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
cmake==3.31.1
comm==0.2.2
compressed-tensors==0.8.0
contourpy==1.3.1
cupy-cuda12x==12.1.0
cycler==0.12.1
dashscope==1.20.14
datasets==3.1.0
debugpy==1.8.9
decorator==5.1.1
deepspeed==0.14.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
faiss-gpu==1.7.2
fastapi==0.115.5
fastrlock==0.8.2
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
Flask==3.1.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
galore-torch==1.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==3.38.0
gradio_client==1.5.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
interegular==0.3.3
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lark==1.2.2
latex2sympy2 @ file:///mnt/workspace/sunhao/code/MARIO_EVAL/latex2sympy
linkify-it-py==2.0.3
# Editable install with no version control (llamafactory==0.8.4.dev0)
-e /mnt/workspace/sunhao/code/LLaMA-Factory-main
# Editable install with no version control (llmtuner==0.6.1)
-e /mnt/workspace/sunhao/code/LLaMA-Factory-main_0.6.1
llvmlite==0.43.0
lm-format-enforcer==0.10.9
markdown-it-py==2.2.0
MarkupSafe==2.1.5
-e git+https://github.com/MARIO-Math-Reasoning/MARIO_EVAL.git@f7a774c1ed7007b0fbd9c9a535a4a0ca9df67ea0#egg=math_evaluation
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistral_common==1.5.1
modelscope==1.20.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.14.3
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.4.0.dev2
openai==1.55.3
opencv-python-headless==4.10.0.84
orjson==3.10.12
outlines==0.0.34
packaging==24.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post4
Pebble==5.0.7
peft==0.13.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.1.0
pycountry==24.6.1
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qwen-agent==0.0.15
ray==2.39.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
ruff==0.8.1
safehttpx==0.1.1
safetensors==0.4.5
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.41.3
sympy==1.12
termcolor==2.5.0
tiktoken==0.7.0
timeout-decorator==0.5.0
tokenizers==0.15.2
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.39.1
triton==2.1.0
trl==0.8.1
typeguard==4.4.1
typer==0.14.0
typing_extensions==4.12.2
tyro==0.9.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
-e git+https://github.com/MARIO-Math-Reasoning/vllm.git@0447598bcab2ae09902568cb92a0bc591811877c#egg=vllm
wandb==0.18.7
watchfiles==1.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
xformers==0.0.23.post1
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
